Week 4 - deleted some files to clean repository
